Some fixes to operand code to remove uneccessary memory/register reads to determine operand sizes.
Change to the handling of RIP-relative addressing to instead use a constant
value for the address of the next instruction. Not sure this is a preferable
approach, but it makes it easier for some analysis tools.

Change to the arguments of the undef instruction to match the behaviour of
BinNavi